[Decomp][Character] Cohesion split: extract customization/equipment handlers to CharacterHandlerCustomize.cpp
